+++ mwdb/model/__init__.py
@@ -5,13 +5,13 @@ db = SQLAlchemy()
 # These imports must appear after "db" declaration
 
 from .api_key import APIKey  # noqa: E402
+from .attribute import Attribute, AttributeDefinition, AttributePermission  # noqa: E402
 from .blob import TextBlob  # noqa: E402
 from .comment import Comment  # noqa: E402
 from .config import Config, StaticConfig  # noqa: E402
 from .file import File  # noqa: E402
 from .group import Group, Member  # noqa: E402
 from .karton import KartonAnalysis, karton_object  # noqa: E402
-from .metakey import Attribute, AttributeDefinition, AttributePermission  # noqa: E402
 from .object import Object, ObjectPermission, relation  # noqa: E402
 from .quick_query import QuickQuery  # noqa: E402
 from .tag import Tag, object_tag_table  # noqa: E402
